requirements.txt
boto3
requests
datetime
pandas
apache-airflow
unidecode
pyarrow
awswrangler
snowflake
TIME-python
